Deep Leaning projects/CNN/project-2(Tuberculosis Detection )/requirements.txt
Flask
tensorflow
pydicom
numpy
opencv-python
pillow
werkzeug
